+++ modules/farm_rothamsted_researcher/src/RothamstedResearcherListBuilder.php
@@ -15,7 +15,6 @@ class RothamstedResearcherListBuilder extends EntityListBuilder {
    */
   public function buildHeader() {
     $header['name'] = $this->t('Name');
-    $header['role'] = $this->t('Role');
     $header['organization'] = $this->t('Organisation');
     $header['department'] = $this->t('Department');
     return $header + parent::buildHeader();
@@ -27,7 +26,6 @@ class RothamstedResearcherListBuilder extends EntityListBuilder {
   public function buildRow(EntityInterface $entity) {
     /** @var \Drupal\asset\Entity\AssetInterface $entity */
     $row['name'] = $entity->toLink($entity->label(), 'canonical')->toString();
-    $row['role'] = $entity->get('role')->value;
     $row['organization'] = $entity->get('organization')->value;
     $row['department'] = $entity->get('department')->value;
     return $row + parent::buildRow($entity);